pyproject.toml
[tool.poetry]
name = "book-review-backend"
version = "0.1.0"
description = "Book Review Platform Backend"
authors = ["Shiv S <shiv.singh@talentica.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
sqlalchemy = ">=2.0.43,<3.0.0"
alembic = ">=1.16.5,<2.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-jose = {extras = ["cryptography"], version = ">=3.5.0,<4.0.0"}
passlib = {extras = ["bcrypt"], version = ">=1.7.4,<2.0.0"}
python-multipart = ">=0.0.20,<0.0.21"
httpx = ">=0.28.1,<0.29.0"
pydantic-settings = ">=2.10.1,<3.0.0"
pillow = ">=11.3.0,<12.0.0"
pytest = ">=8.4.1,<9.0.0"
pytest-asyncio = ">=1.1.0,<2.0.0"
pydantic = {extras = ["email"], version = ">=2.11.7,<3.0.0"}
pytest-aiohttp = ">=1.1.0,<2.0.0"
email-validator = "^2.3.0"
requests = "^2.32.5"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pytest-asyncio = "^1.1.0"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning:passlib.utils"]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
